updated styles
Moved background to body so that it can extend the full visible page without having to make a div for the background. An element would still be needed to allow a background to have a border unless there is another way. I will worry about that if I ever decided to include a title, footer, or nav bar.